Renames package.
